# docs/swagger.yaml
definitions:
  dto.CreateAccountDTO:
    properties:
      currencyId:
        type: integer
      initialAmount:
        type: integer
      name:
        type: string
    required:
    - currencyId
    - initialAmount
    - name
    type: object
  dto.CreateCategoryDTO:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.CreateExpenseDTO:
    properties:
      amount:
        type: integer
      description:
        type: string
      name:
        type: string
    required:
    - amount
    - name
    type: object
  dto.LogInDTO:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterUserDTO:
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - fullName
    - password
    type: object
  dto.UpdateAccountDTO:
    properties:
      currencyId:
        type: integer
      initialAmount:
        type: integer
      name:
        type: string
    required:
    - currencyId
    - initialAmount
    - name
    type: object
  dto.UpdateExpenseDTO:
    properties:
      amount:
        type: integer
      description:
        type: string
      name:
        type: string
    required:
    - amount
    - name
    type: object
  dto.UpdateUserDTO:
    properties:
      email:
        type: string
      fullName:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - fullName
    - password
    type: object
  response.CommonAccountResponse:
    properties:
      currencyId:
        type: integer
      id:
        type: integer
      initialAmount:
        type: integer
      name:
        type: string
      userId:
        type: integer
    type: object
  response.CommonCategoryResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  response.CommonExpenseResponse:
    properties:
      accountId:
        type: integer
      amount:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      userId:
        type: integer
    type: object
  response.CommonUserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: integer
      password:
        type: string
      updatedAt:
        type: string
    type: object
  response.LogInResponse:
    properties:
      token:
        type: string
    type: object
  util.BaseResponse-any:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  util.BaseResponse-array_response_CommonAccountResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.CommonAccountResponse'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  util.BaseResponse-array_response_CommonCategoryResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.CommonCategoryResponse'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  util.BaseResponse-array_response_CommonExpenseResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/response.CommonExpenseResponse'
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
  util.BaseResponse-response_CommonAccountResponse:
    properties:
      data:
        $ref: '#/definitions/response.CommonAccountResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  util.BaseResponse-response_CommonCategoryResponse:
    properties:
      data:
        $ref: '#/definitions/response.CommonCategoryResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  util.BaseResponse-response_CommonExpenseResponse:
    properties:
      data:
        $ref: '#/definitions/response.CommonExpenseResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  util.BaseResponse-response_CommonUserResponse:
    properties:
      data:
        $ref: '#/definitions/response.CommonUserResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: API for Spendtracker
  title: Spendtracker API
  version: "1.0"
paths:
  /accounts:
    post:
      parameters:
      - description: Create account DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAccountDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonAccountResponse'
      summary: Create to account
      tags:
      - account
  /accounts/:
    get:
      parameters:
      - description: Amount of items per page
        in: query
        name: itemPerPage
        type: string
      - description: Page number
        in: query
        name: page
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-array_response_CommonAccountResponse'
      summary: Get many
      tags:
      - account
  /accounts/{accountID}:
    delete:
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-any'
      summary: Delete account
      tags:
      - account
    get:
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonAccountResponse'
      summary: Get one by ID
      tags:
      - account
    put:
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: string
      - description: Update account DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAccountDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonAccountResponse'
      summary: Update account
      tags:
      - account
  /accounts/{accountID}/expenses:
    post:
      parameters:
      - description: Account ID
        in: path
        name: accountId
        required: true
        type: string
      - description: Create expense DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExpenseDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonExpenseResponse'
      summary: Create expense
      tags:
      - expense
  /auth:
    post:
      parameters:
      - description: log in DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.LogInDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LogInResponse'
      summary: Log in to account
      tags:
      - auth
  /categories:
    get:
      parameters:
      - description: Amount of items per page
        in: query
        name: itemPerPage
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-array_response_CommonCategoryResponse'
      summary: Get many categories
      tags:
      - category
    post:
      parameters:
      - description: Create category DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonCategoryResponse'
      summary: Create category
      tags:
      - category
  /categories/{categoryID}:
    delete:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-any'
      summary: Delete one category by ID
      tags:
      - category
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonCategoryResponse'
      summary: Get one category by ID
      tags:
      - category
  /expenses:
    get:
      parameters:
      - description: Account ID
        in: query
        name: accountId
        type: string
      - description: Category ID
        in: query
        name: categoryId
        type: string
      - description: Amount of items per page
        in: query
        name: itemPerPage
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-array_response_CommonExpenseResponse'
      summary: Get many expenses
      tags:
      - expense
  /expenses/{expenseID}:
    delete:
      parameters:
      - description: Expense ID
        in: path
        name: expenseID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-any'
      summary: Delete one expense by ID
      tags:
      - expense
    get:
      parameters:
      - description: Expense ID
        in: path
        name: expenseID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonExpenseResponse'
      summary: Get one expense by ID
      tags:
      - expense
    put:
      parameters:
      - description: Expense ID
        in: path
        name: expenseID
        required: true
        type: string
      - description: Update expense DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateExpenseDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonExpenseResponse'
      summary: Update expense
      tags:
      - expense
  /users:
    post:
      parameters:
      - description: Create user DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserDTO'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonUserResponse'
      summary: Register user
      tags:
      - user
  /users/{userID}:
    put:
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Update user DTO
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.BaseResponse-response_CommonUserResponse'
      summary: Update user
      tags:
      - user
swagger: "2.0"